created testUtility, testDominion1 and testDominion2
testUtility is a duplicate of the playDominion file with utility functions wrapped around sections of the code so that test scenarios may be created and tested in the testDominion files. Each testDominion files then calls the utility functions in testUtility in order to run the program. One or much of the functions are altered slightly to test various behaviors of the code and to introduce desired bugs.